+++ boray/tests/rpc_handlers.rs
@@ -2,6 +2,7 @@
 
 use std::collections::HashMap;
 use std::net::{IpAddr, Ipv4Addr};
+use std::path::Path;
 use std::process::Command;
 use std::sync::Mutex;
 
@@ -11,7 +12,9 @@ use uuid::Uuid;
 
 use cueball::connection_pool::types::ConnectionPoolOptions;
 use cueball::connection_pool::ConnectionPool;
-use cueball_postgres_connection::{PostgresConnection, PostgresConnectionConfig, TlsConfig};
+use cueball_postgres_connection::{
+    PostgresConnection, PostgresConnectionConfig, TlsConfig,
+};
 use cueball_static_resolver::StaticIpResolver;
 use rust_fast::protocol::{FastMessage, FastMessageData};
 
@@ -48,9 +51,10 @@ fn verify_rpc_handlers() {
     // Create pg_tmp database. This requires that pg_tmp be installed on the
     // system running the test.
     ////////////////////////////////////////////////////////////////////////////
-    let create_db_output = Command::new("../tools/postgres/create-ephemeral-db.sh")
-        .output()
-        .expect("failed to execute process");
+    let create_db_output =
+        Command::new("../tools/postgres/create-ephemeral-db.sh")
+            .output()
+            .expect("failed to execute process");
 
     assert!(create_db_output.status.success());
 
@@ -58,7 +62,8 @@ fn verify_rpc_handlers() {
 
     info!(log, "pg url: {}", pg_connect_str);
 
-    let pg_url = Url::parse(&pg_connect_str).expect("failed to parse postgres connection string");
+    let pg_url = Url::parse(&pg_connect_str)
+        .expect("failed to parse postgres connection string");
 
     ////////////////////////////////////////////////////////////////////////////
     // Create connection pool
@@ -100,6 +105,7 @@ fn verify_rpc_handlers() {
     ////////////////////////////////////////////////////////////////////////////
 
     let template_dir = "../schema_templates";
+    let migrations_dir = Path::new("../migrations");
 
     let mut conn = pool
         .claim()
@@ -118,6 +124,7 @@ fn verify_rpc_handlers() {
         &config,
         vnode_resolver,
         template_dir,
+        migrations_dir,
         ["0", "1"].to_vec(),
         &log,
     )
@@ -141,10 +148,14 @@ fn verify_rpc_handlers() {
         request_id,
     };
 
-    let get_bucket_json = serde_json::to_value(vec![get_bucket_payload]).unwrap();
-    let get_bucket_fast_msg_data = FastMessageData::new("getbucket".into(), get_bucket_json);
-    let get_bucket_fast_msg = FastMessage::data(msg_id, get_bucket_fast_msg_data);
-    let mut get_bucket_result = util::handle_msg(&get_bucket_fast_msg, &pool, &log);
+    let get_bucket_json =
+        serde_json::to_value(vec![get_bucket_payload]).unwrap();
+    let get_bucket_fast_msg_data =
+        FastMessageData::new("getbucket".into(), get_bucket_json);
+    let get_bucket_fast_msg =
+        FastMessage::data(msg_id, get_bucket_fast_msg_data);
+    let mut get_bucket_result =
+        util::handle_msg(&get_bucket_fast_msg, &pool, &log);
 
     assert!(get_bucket_result.is_ok());
     let get_bucket_response = get_bucket_result.unwrap();
@@ -166,11 +177,14 @@ fn verify_rpc_handlers() {
         request_id,
     };
 
-    let create_bucket_json = serde_json::to_value(vec![create_bucket_payload]).unwrap();
+    let create_bucket_json =
+        serde_json::to_value(vec![create_bucket_payload]).unwrap();
     let create_bucket_fast_msg_data =
         FastMessageData::new("createbucket".into(), create_bucket_json);
-    let create_bucket_fast_msg = FastMessage::data(msg_id, create_bucket_fast_msg_data);
-    let mut create_bucket_result = util::handle_msg(&create_bucket_fast_msg, &pool, &log);
+    let create_bucket_fast_msg =
+        FastMessage::data(msg_id, create_bucket_fast_msg_data);
+    let mut create_bucket_result =
+        util::handle_msg(&create_bucket_fast_msg, &pool, &log);
 
     assert!(create_bucket_result.is_ok());
     let create_bucket_response = create_bucket_result.unwrap();
@@ -195,7 +209,8 @@ fn verify_rpc_handlers() {
 
     // Try to create same bucket again and verify a BucketAlreadyExists error is
     // returned
-    create_bucket_result = util::handle_msg(&create_bucket_fast_msg, &pool, &log);
+    create_bucket_result =
+        util::handle_msg(&create_bucket_fast_msg, &pool, &log);
 
     assert!(create_bucket_result.is_ok());
     let create_bucket_response = create_bucket_result.unwrap();
@@ -217,11 +232,14 @@ fn verify_rpc_handlers() {
         vnode: 0,
         request_id,
     };
-    let delete_bucket_json = serde_json::to_value(vec![delete_bucket_payload]).unwrap();
+    let delete_bucket_json =
+        serde_json::to_value(vec![delete_bucket_payload]).unwrap();
     let delete_bucket_fast_msg_data =
         FastMessageData::new("deletebucket".into(), delete_bucket_json);
-    let delete_bucket_fast_msg = FastMessage::data(msg_id, delete_bucket_fast_msg_data);
-    let mut delete_bucket_result = util::handle_msg(&delete_bucket_fast_msg, &pool, &log);
+    let delete_bucket_fast_msg =
+        FastMessage::data(msg_id, delete_bucket_fast_msg_data);
+    let mut delete_bucket_result =
+        util::handle_msg(&delete_bucket_fast_msg, &pool, &log);
 
     assert!(delete_bucket_result.is_ok());
     let delete_bucket_response = delete_bucket_result.unwrap();
@@ -248,7 +266,8 @@ fn verify_rpc_handlers() {
     );
 
     // Attempt to delete a nonexistent bucket and verify an error is returned
-    delete_bucket_result = util::handle_msg(&delete_bucket_fast_msg, &pool, &log);
+    delete_bucket_result =
+        util::handle_msg(&delete_bucket_fast_msg, &pool, &log);
 
     assert!(delete_bucket_result.is_ok());
     let delete_bucket_response = delete_bucket_result.unwrap();
@@ -273,10 +292,14 @@ fn verify_rpc_handlers() {
         request_id,
     };
 
-    let get_object_json = serde_json::to_value(vec![&get_object_payload]).unwrap();
-    let get_object_fast_msg_data = FastMessageData::new("getobject".into(), get_object_json);
-    let get_object_fast_msg = FastMessage::data(msg_id, get_object_fast_msg_data);
-    let mut get_object_result = util::handle_msg(&get_object_fast_msg, &pool, &log);
+    let get_object_json =
+        serde_json::to_value(vec![&get_object_payload]).unwrap();
+    let get_object_fast_msg_data =
+        FastMessageData::new("getobject".into(), get_object_json);
+    let get_object_fast_msg =
+        FastMessage::data(msg_id, get_object_fast_msg_data);
+    let mut get_object_result =
+        util::handle_msg(&get_object_fast_msg, &pool, &log);
 
     assert!(get_object_result.is_ok());
     let get_object_response = get_object_result.unwrap();
@@ -315,11 +338,14 @@ fn verify_rpc_handlers() {
         request_id,
     };
 
-    let update_object_json = serde_json::to_value(vec![update_object_payload]).unwrap();
+    let update_object_json =
+        serde_json::to_value(vec![update_object_payload]).unwrap();
     let update_object_fast_msg_data =
         FastMessageData::new("updateobject".into(), update_object_json);
-    let update_object_fast_msg = FastMessage::data(msg_id, update_object_fast_msg_data);
-    let mut update_object_result = util::handle_msg(&update_object_fast_msg, &pool, &log);
+    let update_object_fast_msg =
+        FastMessage::data(msg_id, update_object_fast_msg_data);
+    let mut update_object_result =
+        util::handle_msg(&update_object_fast_msg, &pool, &log);
 
     assert!(update_object_result.is_ok());
     let mut update_object_response = update_object_result.unwrap();
@@ -358,11 +384,14 @@ fn verify_rpc_handlers() {
         request_id,
     };
 
-    let create_object_json = serde_json::to_value(vec![create_object_payload]).unwrap();
+    let create_object_json =
+        serde_json::to_value(vec![create_object_payload]).unwrap();
     let create_object_fast_msg_data =
         FastMessageData::new("createobject".into(), create_object_json);
-    let create_object_fast_msg = FastMessage::data(msg_id, create_object_fast_msg_data);
-    let mut create_object_result = util::handle_msg(&create_object_fast_msg, &pool, &log);
+    let create_object_fast_msg =
+        FastMessage::data(msg_id, create_object_fast_msg_data);
+    let mut create_object_result =
+        util::handle_msg(&create_object_fast_msg, &pool, &log);
 
     assert!(create_object_result.is_ok());
     let create_object_response = create_object_result.unwrap();
@@ -388,7 +417,8 @@ fn verify_rpc_handlers() {
     assert_eq!(&get_object_unwrapped_result.content_type, "text/plain");
 
     // Update the object's metadata and verify it is successful
-    update_object_result = util::handle_msg(&update_object_fast_msg, &pool, &log);
+    update_object_result =
+        util::handle_msg(&update_object_fast_msg, &pool, &log);
 
     assert!(update_object_result.is_ok());
     update_object_response = update_object_result.unwrap();
@@ -418,18 +448,23 @@ fn verify_rpc_handlers() {
 
     // The get and delete object args are the same so we can reuse
     // get_object_json here.
-    let delete_object_json = serde_json::to_value(vec![get_object_payload]).unwrap();
+    let delete_object_json =
+        serde_json::to_value(vec![get_object_payload]).unwrap();
     let delete_object_fast_msg_data =
         FastMessageData::new("deleteobject".into(), delete_object_json);
-    let delete_object_fast_msg = FastMessage::data(msg_id, delete_object_fast_msg_data);
-    let mut delete_object_result = util::handle_msg(&delete_object_fast_msg, &pool, &log);
+    let delete_object_fast_msg =
+        FastMessage::data(msg_id, delete_object_fast_msg_data);
+    let mut delete_object_result =
+        util::handle_msg(&delete_object_fast_msg, &pool, &log);
 
     assert!(delete_object_result.is_ok());
     let delete_object_response = delete_object_result.unwrap();
     assert_eq!(delete_object_response.len(), 1);
 
-    let delete_object_response_result: Result<Vec<object::DeleteObjectResponse>, _> =
-        serde_json::from_value(delete_object_response[0].data.d[0].clone());
+    let delete_object_response_result: Result<
+        Vec<object::DeleteObjectResponse>,
+        _,
+    > = serde_json::from_value(delete_object_response[0].data.d[0].clone());
     assert!(delete_object_response_result.is_ok());
     let delete_object_response = delete_object_response_result.unwrap();
     assert_eq!(delete_object_response.len(), 1);
@@ -453,7 +488,8 @@ fn verify_rpc_handlers() {
     );
 
     // Delete the object again and verify it is not found
-    delete_object_result = util::handle_msg(&delete_object_fast_msg, &pool, &log);
+    delete_object_result =
+        util::handle_msg(&delete_object_fast_msg, &pool, &log);
 
     assert!(delete_object_result.is_ok());
     let delete_object_response = delete_object_result.unwrap();
@@ -478,17 +514,22 @@ fn verify_rpc_handlers() {
         request_id,
     };
 
-    let list_buckets_json = serde_json::to_value(vec![list_buckets_payload]).unwrap();
-    let list_buckets_fast_msg_data = FastMessageData::new("listbuckets".into(), list_buckets_json);
-    let list_buckets_fast_msg = FastMessage::data(msg_id, list_buckets_fast_msg_data);
-    let mut list_buckets_result = util::handle_msg(&list_buckets_fast_msg, &pool, &log);
+    let list_buckets_json =
+        serde_json::to_value(vec![list_buckets_payload]).unwrap();
+    let list_buckets_fast_msg_data =
+        FastMessageData::new("listbuckets".into(), list_buckets_json);
+    let list_buckets_fast_msg =
+        FastMessage::data(msg_id, list_buckets_fast_msg_data);
+    let mut list_buckets_result =
+        util::handle_msg(&list_buckets_fast_msg, &pool, &log);
 
     assert!(list_buckets_result.is_ok());
     let list_buckets_response = list_buckets_result.unwrap();
     assert_eq!(list_buckets_response.len(), 0);
 
     // Create a bucket and list buckets again
-    create_bucket_result = util::handle_msg(&create_bucket_fast_msg, &pool, &log);
+    create_bucket_result =
+        util::handle_msg(&create_bucket_fast_msg, &pool, &log);
 
     assert!(create_bucket_result.is_ok());
     let create_bucket_response = create_bucket_result.unwrap();
@@ -517,17 +558,22 @@ fn verify_rpc_handlers() {
         request_id,
     };
 
-    let list_objects_json = serde_json::to_value(vec![list_objects_payload]).unwrap();
-    let list_objects_fast_msg_data = FastMessageData::new("listobjects".into(), list_objects_json);
-    let list_objects_fast_msg = FastMessage::data(msg_id, list_objects_fast_msg_data);
-    let mut list_objects_result = util::handle_msg(&list_objects_fast_msg, &pool, &log);
+    let list_objects_json =
+        serde_json::to_value(vec![list_objects_payload]).unwrap();
+    let list_objects_fast_msg_data =
+        FastMessageData::new("listobjects".into(), list_objects_json);
+    let list_objects_fast_msg =
+        FastMessage::data(msg_id, list_objects_fast_msg_data);
+    let mut list_objects_result =
+        util::handle_msg(&list_objects_fast_msg, &pool, &log);
 
     assert!(list_objects_result.is_ok());
     let list_objects_response = list_objects_result.unwrap();
     assert_eq!(list_objects_response.len(), 0);
 
     // Create an object and list objects again
-    create_object_result = util::handle_msg(&create_object_fast_msg, &pool, &log);
+    create_object_result =
+        util::handle_msg(&create_object_fast_msg, &pool, &log);
 
     assert!(create_object_result.is_ok());
     let create_object_response = create_object_result.unwrap();
